Switched string names for tokens and ASTNodes to enums instead of strings
Also, changed makefile so it will create a header file containing
variables with file data from the resources folder. This folder is for
files I need access to during compilation that I want to package with
the executable. Examples include: base classes for HexMage variables
for use in the output code, and template files for C++ classes that are
generated by the compiler